Reorder things a bit
Ray pointed out that there is a tiny race in creating the
impl request proxy after the backend call: The backend may
already return the response signal between the backend call
returning and us subscribing to the signal.

Thankfully, we have the id already, so we can create the
impl request proxy ahead of time and avoid the race.